alternate encoding for extension without implicits
